Add proxy_wasm::main macro.
This is primarly a mechanism to deliver a workaround for a breaking
change in Rust v1.56.0 (which updated LLVM to v13.0), which changed
existing reactors into multi-entry commands, leading to performance
degradation and/or unusable Proxy-Wasm plugins.

See: https://github.com/WebAssembly/WASI/issues/471

This is delivered as a macro to make it somehow compatible with the
unstable "-Z wasi-exec-model=reactor" feature.

Signed-off-by: Piotr Sikora <piotrsikora@google.com>